\name{geonames-package}
\alias{geonames-package}
\alias{geonames}
\docType{package}
\title{
R Interface to the www.geonames.org web query service
}
\description{
  This package wraps calls to the web service on www.geonames.org, to
  provide a number of functions for geographic information queries. It
  can query administrative regions, time zones, weather info, cities,
  wikipedia and so on. For more documentation go to
  http://www.geonames.org/ and check the web service info list.

  Note that by default the package uses www.geonames.org as the host
  name. In the past the service has occasionally changed its host name,
  so in order to change the host name used by the package you can set it
  using options(geonamesHost="www2.geonames.org") in R.
}
\details{
\tabular{ll}{
Package: \tab geonames\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2008-10-10\cr
License: \tab GPL v3\cr
LazyLoad: \tab yes\cr
}

}
\author{
Barry Rowlingson

Maintainer: Barry Rowlingson <barry.rowlingson@gmail.com>

}
\keyword{ package }

