% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonames.R
\docType{package}
\name{geonames}
\alias{geonames}
\alias{geonames-package}
\title{Query the geonames web API for geographic data}
\description{
www.geonames.org is a service where you can query for global geographic
data such as administrative areas, populated places, weather data etc.
}
\details{
The functions in this package are mostly thin wrappers to the API
calls documented at the geonames web services
overview \url{http://www.geonames.org/export/ws-overview.html}.

A set of example calls are supplied in a file with the package. Once
you have set your geonames username with
\code{options(geonamesUsername="myusernamehere")} you can run
this with \code{source(system.file("tests","testing.R",package="geonames"),echo=TRUE)}
}
