% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GNcities}
\alias{GNcities}
\title{find cities}
\usage{
GNcities(north, east, south, west, lang = "en", maxRows = 10)
}
\arguments{
\item{north}{north bound}

\item{east}{east bound}

\item{south}{south bound}

\item{west}{west bount}

\item{lang}{language code}

\item{maxRows}{max number of records to return}
}
\value{
city records
}
\description{
Search for cities
}
\details{
find cities

API doc for GNcities is at \url{http://www.geonames.org/export/JSON-webservices.html#citiesJSON}
}
\author{
Barry Rowlingson
}

