% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{GNcountryInfo}
\alias{GNcountryInfo}
\title{country info}
\usage{
GNcountryInfo(country = "", lang = "")
}
\arguments{
\item{country}{country code}

\item{lang}{language code}
}
\value{
country record info
}
\description{
Get country info
}
\details{
country info

API doc for GNcountryInfo is at \url{http://www.geonames.org/export/web-services.html#countryInfo}
}
\author{
Barry Rowlingson
}
