% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNcountrySubdivision}
\alias{GNcountrySubdivision}
\title{country code and subdivision}
\usage{
GNcountrySubdivision(lat, lng, lang = "en", radius = "", maxRows = 10)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{lang}{language code}

\item{radius}{search radius}

\item{maxRows}{max number of returned records}
}
\value{
iso country code
}
\description{
country code and admin subdivision
}
\details{
looks up country and admin subdivisions
}
\author{
Barry Rowlingson
}

