% Generated by roxygen2 (4.0.0): do not edit by hand
\name{GNearthquakes}
\alias{GNearthquakes}
\title{recent earthquakes}
\usage{
GNearthquakes(north, east, south, west, date, minMagnitude, maxRows = 10)
}
\arguments{
\item{north}{north bound}

\item{east}{east bound}

\item{south}{south bound}

\item{west}{west bound}

\item{date}{optional date}

\item{minMagnitude}{optional minimal magnitude}

\item{maxRows}{max records to return}
}
\value{
earthquake records
}
\description{
recent earthquakes
}
\details{
get recent earthquakes in a region

API doc for GNearthquakes is at \url{http://www.geonames.org/export/JSON-webservices.html#earthquakesJSON}
}
\examples{
\dontrun{
GNearthquakes(north=44.1,south=-9.9,east=-22.4,west=55.2)
}
}
\author{
Barry Rowlingson
}

