% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{GNfindNearby}
\alias{GNfindNearby}
\title{nearby search}
\usage{
GNfindNearby(...)
}
\arguments{
\item{...}{search parameters}
}
\value{
matched records
}
\description{
find nearby entities
}
\details{
nearby search

API doc for GNfindNearby is at \url{http://www.geonames.org/export/web-services.html#findNearby}
}
\author{
Barry Rowlingson
}
