% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GNfindNearbyPostalCodes}
\alias{GNfindNearbyPostalCodes}
\title{find postal code}
\usage{
GNfindNearbyPostalCodes(...)
}
\arguments{
\item{...}{search parameters, see geonames web docs for details}
}
\value{
postal code records
}
\description{
find postal code by lat long or code
}
\details{
find postal code

API doc for GNfindNearbyPostalCodes is at \url{http://www.geonames.org/export/web-services.html#findNearbyPostalCodes}
}
\author{
Barry Rowlingson
}

