% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNfindNearByWeather}
\alias{GNfindNearByWeather}
\title{weather at location}
\usage{
GNfindNearByWeather(lat, lng)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}
}
\value{
weather record
}
\description{
get weather at location
}
\details{
get weather
}
\note{
check capitalisation of 'NearBy'
}
\author{
Barry Rowlingson
}

