% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia.R
\name{GNfindNearbyWikipedia}
\alias{GNfindNearbyWikipedia}
\title{nearby wikipedia entries}
\usage{
GNfindNearbyWikipedia(...)
}
\arguments{
\item{...}{see geonames.org documentation}
}
\value{
wikipedia entries
}
\description{
find nearby wikipedia entries
}
\details{
search wikipedia entries by lat/lng or location name parameters

API doc for GNfindNearbyWikipedia is at \url{http://www.geonames.org/export/wikipedia-webservice.html#findNearbyWikipedia}
}
\examples{
\dontrun{
GNfindNearbyWikipedia(postalcode=8775,country="CH",radius=10)
}
}
\author{
Barry Rowlingson
}
