% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{GNfindNearestAddress}
\alias{GNfindNearestAddress}
\title{nearest address}
\usage{
GNfindNearestAddress(lat, lng)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}
}
\value{
address record
}
\description{
find nearest street and address
}
\details{
search US for nearest street and address

API doc for GNfindNearestAddress is at \url{http://www.geonames.org/maps/us-reverse-geocoder.html#findNearestAddress}
}
\author{
Barry Rowlingson
}
