% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNgtopo30}
\alias{GNgtopo30}
\title{topo30 height}
\usage{
GNgtopo30(lat, lng)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}
}
\value{
height record
}
\description{
height from topo30
}
\details{
get height from topo30 data
}
\author{
Barry Rowlingson
}

