% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{GNneighbourhood}
\alias{GNneighbourhood}
\title{neighbourhood}
\usage{
GNneighbourhood(lat, lng)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}
}
\value{
neighbourhood records
}
\description{
find neighbourhood
}
\details{
find neighbourhood

API doc for GNneighbourhood is at \url{http://www.geonames.org/export/web-services.html#neighbourhood}
}
\author{
Barry Rowlingson
}
