% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNpostalCodeCountryInfo}
\alias{GNpostalCodeCountryInfo}
\title{postal code info}
\usage{
GNpostalCodeCountryInfo()
}
\value{
list of countries with postal codes on geonames
}
\description{
countries with postal code info
}
\details{
list countries with postal code info
}
\author{
Barry Rowlingson
}

