% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal.R
\name{GNpostalCodeLookup}
\alias{GNpostalCodeLookup}
\title{postal code lookup}
\usage{
GNpostalCodeLookup(...)
}
\arguments{
\item{...}{parameters}
}
\value{
list of places for a given input postal code
}
\description{
postal code lookup
}
\details{
postal code lookup

API doc for GNpostalCodeLookup is at \url{http://www.geonames.org/export/web-services.html#postalCodeLookupJSON}
}
\author{
Barry Rowlingson
}
