% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal.R
\name{GNpostalCodeSearch}
\alias{GNpostalCodeSearch}
\title{postal code search}
\usage{
GNpostalCodeSearch(...)
}
\arguments{
\item{...}{search parameters}
}
\value{
postal code record
}
\description{
search for postal code
}
\details{
full text search for postal codes

API doc for GNpostalCodeSearch is at \url{http://www.geonames.org/export/web-services.html#postalCodeSearch}
}
\author{
Barry Rowlingson
}
