% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{GNsearch}
\alias{GNsearch}
\title{search geonames}
\usage{
GNsearch(...)
}
\arguments{
\item{...}{search parameters}
}
\value{
matched records
}
\description{
search geonames
}
\details{
general search call

API doc for GNsearch is at \url{http://www.geonames.org/export/geonames-search.html}
}
\examples{
\dontrun{
# Find places called Lancaster and 'geocode' the result
(lanc_df <- GNsearch(name = "Lancaster", country = "UK"))
lanc_coords <- lanc_df[1, c("lng", "lat")]
}
}
\author{
Barry Rowlingson
}
