% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{GNtimezone}
\alias{GNtimezone}
\title{timezone for location}
\usage{
GNtimezone(lat, lng, radius = 0)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{radius}{sesarch radius}
}
\value{
time zone record
}
\description{
get timezone
}
\details{
timezone for location

API doc for GNtimezone is at \url{http://www.geonames.org/export/web-services.html#timezone}
}
\author{
Barry Rowlingson
}
