% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNweather}
\alias{GNweather}
\title{weather stations in box}
\usage{
GNweather(north, east, south, west, maxRows = 10)
}
\arguments{
\item{north}{north bound}

\item{east}{east bound}

\item{south}{south bound}

\item{west}{west bound}

\item{maxRows}{max records to return}
}
\value{
weather records
}
\description{
weather stations in region
}
\details{
get weather stations in region with latest readings
}
\author{
Barry Rowlingson
}

