% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNweatherIcao}
\alias{GNweatherIcao}
\title{ICAO weather station data}
\usage{
GNweatherIcao(ICAO)
}
\arguments{
\item{ICAO}{ICAO code}
}
\value{
weather record
}
\description{
weather record from ICAO station
}
\details{
get most recent ICAO station data
}
\author{
Barry Rowlingson
}

