% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNwikipediaBoundingBox}
\alias{GNwikipediaBoundingBox}
\title{wikipedia articles in a box}
\usage{
GNwikipediaBoundingBox(...)
}
\arguments{
\item{...}{parameters (north, south, east, west etc.)}
}
\value{
wikipedia records
}
\description{
wikipedia articles in bounding box
}
\details{
find articles in a box

API doc for GNwikipediaBoundingBox is at \url{http://www.geonames.org/export/wikipedia-webservice.html#wikipediaBoundingBox}
}
\author{
Barry Rowlingson
}

