% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNwikipediaSearch}
\alias{GNwikipediaSearch}
\title{search wikipedia}
\usage{
GNwikipediaSearch(q, maxRows = 10)
}
\arguments{
\item{q}{search string}

\item{maxRows}{maximum returned records}
}
\value{
wikipedia entries
}
\description{
wikipedia fulltext search
}
\details{
find geolocated articles in wikipedia
}
\author{
Barry Rowlingson
}

