% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{hierarchy}
\alias{GNchildren}
\alias{GNhierarchy}
\alias{GNneighbours}
\alias{GNsiblings}
\alias{hierarchy}
\title{Admin area hierarchy}
\usage{
GNchildren(geonameId, ...)

GNhierarchy(geonameId, ...)

GNsiblings(geonameId, ...)

GNneighbours(geonameId, ...)
}
\arguments{
\item{geonameId}{a geonames ID value}

\item{...}{other parameters to pass to geonames}
}
\description{
See \url{http://www.geonames.org/export/ws-overview.html}
for a full description of valid arguments and return values
}
\details{
API doc for GNchildren is at \url{http://www.geonames.org/export/place-hierarchy.html#children}

API doc for GNhierarchy is at \url{http://www.geonames.org/export/place-hierarchy.html#hierarchy}

API doc for GNsiblings is at \url{http://www.geonames.org/export/place-hierarchy.html#siblings}

API doc for GNneighbours is at \url{http://www.geonames.org/export/place-hierarchy.html#neighbours}
}

