% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNcountryCode}
\alias{GNcountryCode}
\title{country code for location}
\usage{
GNcountryCode(lat, lng, lang = "", radius = "")
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{lang}{language code}

\item{radius}{radius size}
}
\value{
country record
}
\description{
Get country code
}
\details{
country code for location
}
\author{
Barry Rowlingson
}

