% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GNcountryCode}
\alias{GNcountryCode}
\title{country code for location}
\usage{
GNcountryCode(lat, lng, lang = "", radius = "")
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{lang}{language code}

\item{radius}{radius size}
}
\value{
country record
}
\description{
Get country code
}
\details{
country code for location

API doc for GNcountryCode is at \url{http://www.geonames.org/export/web-services.html#countrycode}
}
\author{
Barry Rowlingson
}

