% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNcountryInfo}
\alias{GNcountryInfo}
\title{country info}
\usage{
GNcountryInfo(country = "", lang = "")
}
\arguments{
\item{country}{country code}

\item{lang}{language code}
}
\value{
country record info
}
\description{
Get country info
}
\details{
country info
}
\author{
Barry Rowlingson
}

