% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNfindNearby}
\alias{GNfindNearby}
\title{nearby search}
\usage{
GNfindNearby(...)
}
\arguments{
\item{...}{search parameters}
}
\value{
matched records
}
\description{
find nearby entities
}
\details{
nearby search
}
\author{
Barry Rowlingson
}

