% Generated by roxygen2 (4.0.0): do not edit by hand
\name{GNfindNearby}
\alias{GNfindNearby}
\title{nearby search}
\usage{
GNfindNearby(...)
}
\arguments{
\item{...}{search parameters}
}
\value{
matched records
}
\description{
find nearby entities
}
\details{
nearby search

API doc for GNfindNearby is at \url{http://www.geonames.org/export/web-services.html#findNearby}
}
\author{
Barry Rowlingson
}

