% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNfindNearbyStreets}
\alias{GNfindNearbyStreets}
\title{nearby street finding}
\usage{
GNfindNearbyStreets(lat, lng)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}
}
\value{
street records
}
\description{
find nearby streets (US only)
}
\details{
for a lat-long, find nearby US streets

API doc for GNfindNearbyStreets is at \url{http://www.geonames.org/maps/us-reverse-geocoder.html#findNearbyStreets}
}
\author{
Barry Rowlingson
}

