% Generated by roxygen2 (4.0.0): do not edit by hand
\name{GNfindNearByWeather}
\alias{GNfindNearByWeather}
\title{weather at location}
\usage{
GNfindNearByWeather(lat, lng)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}
}
\value{
weather record
}
\description{
get weather at location
}
\details{
get weather

API doc for GNfindNearByWeather is at \url{http://www.geonames.org/export/JSON-webservices.html#findNearByWeatherJSON}
}
\note{
check capitalisation of 'NearBy'
}
\examples{
\dontrun{
GNfindNearByWeather(57,-2)
}
}
\author{
Barry Rowlingson
}

