% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNfindNearestAddress}
\alias{GNfindNearestAddress}
\title{nearest address}
\usage{
GNfindNearestAddress(lat, lng)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}
}
\value{
address record
}
\description{
find nearest street and address
}
\details{
search US for nearest street and address
}
\author{
Barry Rowlingson
}

