% Generated by roxygen2 (4.0.0): do not edit by hand
\name{GNfindNearestIntersection}
\alias{GNfindNearestIntersection}
\title{nearest intersection}
\usage{
GNfindNearestIntersection(lat, lng)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}
}
\value{
intersection record
}
\description{
search US for nearest intersection
}
\details{
finds nearest intersection

API doc for GNfindNearestIntersection is at \url{http://www.geonames.org/maps/us-reverse-geocoder.html#findNearestIntersection}
}
\author{
Barry Rowlingson
}

