% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topography.R
\name{GNgtopo30}
\alias{GNgtopo30}
\title{topo30 height}
\usage{
GNgtopo30(lat, lng)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}
}
\value{
height record
}
\description{
height from topo30
}
\details{
get height from topo30 data

API doc for GNgtopo30 is at \url{http://www.geonames.org/export/web-services.html#gtopo30}
}
\examples{
\dontrun{
GNgtopo30(lat=54,lng=-1)
}
}
\author{
Barry Rowlingson
}
