% Generated by roxygen2 (4.0.0): do not edit by hand
\name{GNneighbourhood}
\alias{GNneighbourhood}
\title{neighbourhood}
\usage{
GNneighbourhood(lat, lng)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}
}
\value{
neighbourhood records
}
\description{
find neighbourhood
}
\details{
find neighbourhood

API doc for GNneighbourhood is at \url{http://www.geonames.org/export/web-services.html#neighbourhood}
}
\author{
Barry Rowlingson
}

