% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNneighbourhood}
\alias{GNneighbourhood}
\title{neighbourhood}
\usage{
GNneighbourhood(lat, lng)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}
}
\value{
neighbourhood records
}
\description{
find neighbourhood
}
\details{
find neighbourhood
}
\author{
Barry Rowlingson
}

