% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal.R
\name{GNpostalCodeCountryInfo}
\alias{GNpostalCodeCountryInfo}
\title{postal code info}
\usage{
GNpostalCodeCountryInfo()
}
\value{
list of countries with postal codes on geonames
}
\description{
countries with postal code info
}
\details{
list countries with postal code info

API doc for GNpostalCodeCountryInfo is at \url{http://www.geonames.org/export/web-services.html#postalCodeCountryInfo}
}
\author{
Barry Rowlingson
}
