% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNpostalCodeLookup}
\alias{GNpostalCodeLookup}
\title{postal code lookup}
\usage{
GNpostalCodeLookup(...)
}
\arguments{
\item{...}{parameters}
}
\value{
list of places for a given input postal code
}
\description{
postal code lookup
}
\details{
postal code lookup
}
\author{
Barry Rowlingson
}

