% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNpostalCodeSearch}
\alias{GNpostalCodeSearch}
\title{postal code search}
\usage{
GNpostalCodeSearch(...)
}
\arguments{
\item{...}{search parameters}
}
\value{
postal code record
}
\description{
search for postal code
}
\details{
full text search for postal codes

API doc for GNpostalCodeSearch is at \url{http://www.geonames.org/export/web-services.html#postalCodeSearch}
}
\author{
Barry Rowlingson
}

