% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNsearch}
\alias{GNsearch}
\title{search geonames}
\usage{
GNsearch(...)
}
\arguments{
\item{...}{search parameters}
}
\value{
matched records
}
\description{
search geonames
}
\details{
general search call
}
\author{
Barry Rowlingson
}

