% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNsrtm3}
\alias{GNsrtm3}
\title{srtm3 height}
\usage{
GNsrtm3(lat, lng)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}
}
\value{
height record
}
\description{
height from srtm3 data
}
\details{
get srtm3 height

API doc for GNsrtm3 is at \url{http://www.geonames.org/export/web-services.html#srtm3}
}
\author{
Barry Rowlingson
}

