% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{GNtimezone}
\alias{GNtimezone}
\title{timezone for location}
\usage{
GNtimezone(lat, lng, radius = 0)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{radius}{sesarch radius}
}
\value{
time zone record
}
\description{
get timezone
}
\details{
timezone for location
}
\author{
Barry Rowlingson
}

