% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GNweatherIcao}
\alias{GNweatherIcao}
\title{ICAO weather station data}
\usage{
GNweatherIcao(ICAO)
}
\arguments{
\item{ICAO}{ICAO code}
}
\value{
weather record
}
\description{
weather record from ICAO station
}
\details{
get most recent ICAO station data

API doc for GNweatherIcao is at \url{http://www.geonames.org/export/JSON-webservices.html#weatherIcaoJSON}
}
\author{
Barry Rowlingson
}

